apino update

+++ src/main/java/kr/co/dajsoft/hell0/dto/ApiDTO.java
@@ -7,7 +7,7 @@ import lombok.*;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ApiDTO {
-    private Long apino;
+    private int apino;
     private String gubun;
     private String maxclassnm;
     private String minclassnm;
